+++ R/board_registration.R
@@ -215,6 +215,7 @@ board_register_datatxt <- function(url,
 #' @param secret The secret of the Amazon S3 bucket. Defaults to the \code{AWS_SECRET_ACCESS_KEY} environment
 #'   variable.
 #' @param cache The local folder to use as a cache, defaults to \code{board_cache_path()}.
+#' @param host The host to use for storage, defaults to \code{"s3.amazonaws.com"}.
 #' @param ... Additional parameters required to initialize a particular board.
 #'
 #' @details
@@ -235,6 +236,7 @@ board_register_s3 <- function(name = "s3",
                               key = Sys.getenv("AWS_ACCESS_KEY_ID"),
                               secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
                               cache = board_cache_path(),
+                              host = "s3.amazonaws.com",
                               ...) {
   board_register("s3",
                  name = name,
@@ -325,5 +327,55 @@ board_register_gcloud <- function(name = "gcloud",
                  bucket = bucket,
                  token = token,
                  cache = cache,
+                 ...)
+}
+
+#' Register DigitalOcean Board
+#'
+#' Wrapper with explicit parameters over \code{board_register()} to
+#' register a DigitalOcean Spaces board.
+#'
+#' @param name Optional name for this board, defaults to 's3'.
+#' @param space The name of the DigitalOcean space. Defaults to the \code{DO_SPACE} environment
+#'   variable.
+#' @param key The key of the DigitalOcean space. Defaults to the \code{DO_ACCESS_KEY_ID} environment
+#'   variable.
+#' @param secret The secret of the DigitalOcean space. Defaults to the \code{DO_SECRET_ACCESS_KEY} environment
+#'   variable.
+#' @param datacenter The datacenter of the DigitalOcean space. Defaults to the \code{DO_DATACENTER} environment
+#'   variable.
+#' @param cache The local folder to use as a cache, defaults to \code{board_cache_path()}.
+#' @param host The host to use for storage, defaults to \code{"digitaloceanspaces.com"}.
+#' @param ... Additional parameters required to initialize a particular board.
+#'
+#' @details
+#'
+#' This function requires a DigitalOcean space to be manually created; otherwise,
+#' registering a DigitalOcean space will fail.
+#'
+#' @seealso board_register
+#'
+#' @examples
+#' \dontrun{
+#' # the following example requires a DigitalOcean Spaces API key
+#' board_register_s3(bucket = "s3bucket")
+#' }
+#' @export
+board_register_dospace <- function(name = "dospace",
+                                   space = Sys.getenv("DO_SPACE"),
+                                   key = Sys.getenv("DO_ACCESS_KEY_ID"),
+                                   secret = Sys.getenv("DO_SECRET_ACCESS_KEY"),
+                                   datacenter = Sys.getenv("DO_DATACENTER"),
+                                   cache = board_cache_path(),
+                                   host = "digitaloceanspaces.com",
+                                   ...) {
+  board_register("dospace",
+                 name = name,
+                 space = space,
+                 key = key,
+                 secret = secret,
+                 datacenter = datacenter,
+                 cache = cache,
+                 host = host,
                  ...)
 }